+++ src/main/java/roomescape/reservation/application/ReservationService.java
@@ -8,8 +8,8 @@ import java.util.NoSuchElementException;
 import java.util.Objects;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import roomescape.admin.reservation.AdminReservationRequest;
-import roomescape.admin.reservation.AdminWaitingReservationResponse;
+import roomescape.admin.reservation.presentation.dto.AdminReservationRequest;
+import roomescape.admin.reservation.presentation.dto.AdminWaitingReservationResponse;
 import roomescape.member.application.MemberRepository;
 import roomescape.member.domain.Member;
 import roomescape.reservation.domain.Reservation;
